Move converse-roomslist plugin into folder
