+++ src/converse.js
@@ -28,7 +28,7 @@ import "./plugins/omemo.js";
 import "./plugins/profile.js";
 import "./plugins/push.js";                 // XEP-0357 Push Notifications
 import "./plugins/register.js";             // XEP-0077 In-band registration
-import "./plugins/roomslist.js";            // Show currently open chat rooms
+import "./plugins/roomslist/index.js";      // Show currently open chat rooms
 import "./plugins/rosterview.js";
 import "./plugins/singleton.js";
 /* END: Removable components */
